fix(transaction-manager): store not work
